segregate variables for browser api
